fix(analysis) #231: 분석 실패 시 그래프 롤백과 상태 보존
