+++ src/main/java/com/saferoute/domain/evacuation/grid/repository/MapEdgeGridCellRepository.java
@@ -18,7 +18,9 @@ public interface MapEdgeGridCellRepository extends JpaRepository<MapEdgeGridCell
     List<MapEdgeGridCell> findAllByMapEdge_Id(UUID mapEdgeId);
 
     // AI 분석으로 층 그래프 전체가 교체된 뒤 기존 매핑만 정리하고 다시 계산할 때 사용
-    void deleteAllByMapEdge_Floor_Id(UUID floorId);
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
+    @Query("delete from MapEdgeGridCell m where m.mapEdge.floor.id = :floorId")
+    int deleteAllByMapEdgeFloorId(@Param("floorId") UUID floorId);
 
     // 엣지 하나의 grid cell 매핑을 다시 계산하기 전, 기존 매핑을 지운다
     // (엣지 추가/노드 이동으로 인한 재계산용 - FloorGridService.recomputeEdgeGridCells 참고).
